backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-firestore
google-cloud-storage
python-dotenv
python-multipart
functions-framework